requirements.txt
pandas
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
requests==2.31.0
yfinance==0.2.35
ta==0.11.0
